django/requirements.txt
Django==2.1.13
django-filter==2.2.0
django-pyodbc-azure==2.1.0.0
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.4
pyodbc==4.0.27
pytz==2019.3
sqlparse==0.3.0
